+++ WidgetExtension/ActivityAttributes.swift
@@ -1,6 +1,6 @@
 import ActivityKit
 
-struct iOSWidgetExtensionAttributes: ActivityAttributes {
+struct WidgetExtensionAttributes: ActivityAttributes {
     public struct ContentState: Codable, Hashable {
         var slot: UInt8
     }
