+++ WidgetExtension/AppIntent.swift
@@ -1,10 +1,3 @@
-//
-//  AppIntent.swift
-//  iOSWidgetExtension
-//
-//  Created by Cameron Little on 2023-12-08.
-//
-
 import WidgetKit
 import AppIntents
