+++ WidgetExtension/Assets.xcassets/AccentColor.colorset/Contents.json
