+++ WidgetExtension/Assets.xcassets/Contents.json
