+++ WidgetExtension/Assets.xcassets/WidgetBackground.colorset/Contents.json
