+++ WidgetExtension/Info.plist
