+++ WidgetExtension/WidgetExtensionBundle.swift
@@ -1,0 +1,9 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct WidgetExtensionBundle: WidgetBundle {
+    var body: some Widget {
+        WidgetExtensionLiveActivity()
+    }
+}
