+++ iOS/ActivityManager.swift
@@ -2,16 +2,16 @@ import ActivityKit
 
 @available(iOS 16.2, *)
 class ActivityManager {
-    static var activity: Activity<iOSWidgetExtensionAttributes>? = nil
+    static var activity: Activity<WidgetExtensionAttributes>? = nil
     
     static func update(slot: UInt8?) async {
         if let slot = slot {
-            let content = ActivityContent(state: iOSWidgetExtensionAttributes.ContentState(slot: slot), staleDate: nil)
+            let content = ActivityContent(state: WidgetExtensionAttributes.ContentState(slot: slot), staleDate: nil)
             if let activeActivity = self.activity {
                 await activeActivity.update(content)
             } else {
                 do {
-                    Self.activity = try Activity.request(attributes: iOSWidgetExtensionAttributes(), content: content)
+                    Self.activity = try Activity.request(attributes: WidgetExtensionAttributes(), content: content)
                 } catch {
                     print("error: \(error)")
                 }
